+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* Allow element in events to be a function that returns an element
+
 ## v0.3.0 - 2025/10/17
 * Added a new `FormAutosubmitElement`
 * Added a `debounce` function (used in the `FormAutosubmitElement` that is also exported)